JS/SDK: Handle relayer move
