+++ sdk/js/scripts/copyContracts.js
@@ -1,3 +1,7 @@
 const copydir = require("copy-dir");
 console.log("Copying from ../../ethereum/ethers-contracts");
 copydir.sync("../../ethereum/ethers-contracts", "src/ethers-contracts");
+copydir.sync(
+  "../../ethereum-relayer/ethers-contracts",
+  "src/ethers-relayer-contracts"
+);
