+++ testing/Dockerfile.sdk.test
@@ -17,6 +17,11 @@ RUN --mount=type=cache,uid=1000,gid=1000,target=/home/node/.npm \
   npm ci --prefix ethereum
 COPY ethereum ./ethereum
 
+COPY ethereum-relayer/package.json ethereum-relayer/package-lock.json ./ethereum-relayer/
+RUN --mount=type=cache,uid=1000,gid=1000,target=/home/node/.npm \
+  npm ci --prefix ethereum-relayer
+COPY ethereum-relayer ./ethereum-relayer
+
 COPY solana/idl ./solana/idl/
 
 COPY sdk/js/package.json sdk/js/package-lock.json  ./sdk/js/
